split `cidrw` into two modules: `cidrw` and `cidobj`; new `CidObjError`
