+++ candejar/cidobj/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Sub package for working with CANDE problems as a Python data model object."""
