+++ candejar/cidobj/cidobj_abc.py
