+++ candejar/cidobj/cidseq/names.py
