+++ candejar/cidobj/cidsubobj/names.py
