+++ candejar/cidobj/exc.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+"""`cidobj` module exceptions"""
+
+class CIDObjError(Exception):
+    """Base exception for `cidobj` module"""
+    pass
+
