+++ candejar/cidobj/names.py
