+++ tests/test_cidobj_abc.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-"""Tests for `candejar.cidrw.cidobj_abc` module."""
+"""Tests for `candejar.cidobj.cidobj_abc` module."""
 import pytest
-from candejar.cidrw.cidobj_abc import CidObjBase, CidABC
+from candejar.cidobj.cidobj_abc import CidObjBase, CidABC
 
 
 def test_cidabc_subclass():
